news_sarathi/requirements.txt
adbs==0.0.2
asgiref==3.4.1
Django==3.2.6
django-admin-honeypot==1.1.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-mysql==3.12.0
django-widget-tweaks==1.4.8
extras==1.0.0
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
